Files Added
These are the foundation files used to create an express js application.

Dependencies included need to be installed using ‘npm install’.

Mongodb and node needs to be installed.

Mongoose is used to interact with mongodb in the application, which is
probably the easiest module I’ve found.

Views include bootstrap and jQuery. Example Views, Routes and DB schema
are included.